+++ test/e2e/change-behavior.spec.js
@@ -8,7 +8,7 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-const { startServer, stopServer, request } = require("./utils");
+const { startServer, stopServer, request } = require("./support/utils");
 
 describe("API for changing current behavior", () => {
   let server;
@@ -17,8 +17,8 @@ describe("API for changing current behavior", () => {
     server = await startServer("web-tutorial-json");
   });
 
-  afterAll(() => {
-    stopServer(server);
+  afterAll(async () => {
+    await stopServer(server);
   });
 
   describe("When started", () => {