+++ test/e2e/deprecated-core-events.spec.js
@@ -17,7 +17,7 @@ const {
   fixturesFolder,
   wait,
   CliRunner,
-} = require("./utils");
+} = require("./support/utils");
 
 describe("Plugin listening to core events", () => {
   let cli;
@@ -28,7 +28,7 @@ describe("Plugin listening to core events", () => {
     cli = new CliRunner(["node", "start-files-watch.js"], {
       cwd: path.resolve(__dirname, "fixtures"),
     });
-    await wait(1000);
+    await wait(2000);
   });
 
   afterAll(async () => {