+++ CHANGELOG.md
@@ -14,6 +14,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [1.5.0] - 2020-12-25
+
+### Added
+- feat(#78): Add alerts router
+- chore(lint): Add eslint plugin to avoid only in tests
+
+### Changed
+- chore(deps): Update dependencies
+- test(e2e): Move utils to a support folder
+- feat: Add/remove routers on start/stop plugin methods, not in init method.
+- style: Remove redundant `await` on a non-promise
+
 ## [1.4.7] - 2020-12-21
 
 ### Added